* Zone.m ([Zone_c -describe:]): Cast all counts to unsigned long and use %lu specifier.
* Zone.h: Use size_t instead of unsigned for return of getPageSize.
(Zone_c): Use size_t for all counts.


git-svn-id: svn://svn.savannah.nongnu.org/swarm/trunk@3628 84d386b0-d280-4341-8980-c4cb7e1eb7a5

+++ swarm/src/defobj/Zone.h
@@ -17,11 +17,11 @@ Library:      defobj
   id componentZone;          // view of zone qualified for component allocation
   id population;             // external objects created within the zone
   size_t populationTotal;    // total size of population objects
-  unsigned objectCount;      // count of internal objects allocated in the zone
+  size_t objectCount;      // count of internal objects allocated in the zone
   size_t objectTotal;        // total size of internal objects
-  unsigned blockCount;       // count of internal blocks allocated in the zone
+  size_t blockCount;       // count of internal blocks allocated in the zone
   size_t blockTotal;         // total size of internal blocks
-  unsigned allocCount;       // count of blocks allocated by alloc:
+  size_t allocCount;       // count of blocks allocated by alloc:
   size_t allocTotal;         // total size of alloc'ed blocks
   id internalAllocations;    // collection of all internal allocations
 }
